+++ src/interfaces/ILiteSequencerClient.ts
@@ -7,7 +7,6 @@ import {
     OperationType,
     StatusInfosByOperationId,
     SuggestedTVMExecutorFee,
-    TacGasPriceResponse,
     TACSimulationParams,
     TACSimulationResult,
     TransactionLinker,
@@ -61,10 +60,4 @@ export interface ILiteSequencerClient {
      * @returns Promise resolving to detailed simulation result.
      */
     simulateTACMessage(params: TACSimulationParams): Promise<TACSimulationResult>;
-
-    /**
-     * Gets TAC gas price.
-     * @returns Promise resolving to TAC gas price.
-     */
-    getTACGasPrice(): Promise<TacGasPriceResponse>;
 }