+++ src/sdk/LiteSequencerClient.ts
@@ -1,7 +1,6 @@
 import { emptyArrayError, operationFetchError, profilingFetchError, statusFetchError } from '../errors';
 import {
     convertCurrencyFetchError,
-    gasPriceFetchError,
     getTONFeeInfoFetchError,
     simulationFetchError,
 } from '../errors/instances';
@@ -26,7 +25,6 @@ import {
     OperationType,
     StatusInfosByOperationId,
     SuggestedTVMExecutorFee,
-    TacGasPriceResponse,
     TACSimulationParams,
     TACSimulationResult,
     TransactionLinker,
@@ -258,18 +256,6 @@ export class LiteSequencerClient implements ILiteSequencerClient {
             return response.data.response;
         } catch (error) {
             throw simulationFetchError(`endpoint ${this.endpoint} failed to complete request`, error);
-        }
-    }
-
-    async getTACGasPrice(): Promise<TacGasPriceResponse> {
-        try {
-            const response = await this.httpClient.get<TacGasPriceResponse>(
-                new URL('stats', 'https://explorer.tac.build/api/v2/').toString(),
-                { transformResponse: [toCamelCaseTransformer] },
-            );
-            return response.data;
-        } catch (error) {
-            throw gasPriceFetchError(`endpoint https://explorer.tac.build/api/v2/ failed to complete request`, error);
         }
     }