+++ src/sdk/TacSdk.ts
@@ -9,6 +9,7 @@ import {
     ILogger,
     IOperationTracker,
     ISimulator,
+    ITacExplorerClient,
     ITacSDK,
     ITACTransactionManager,
     ITONTransactionManager,
@@ -33,6 +34,7 @@ import {
     NFTAddressType,
     SDKParams,
     SuggestedTVMExecutorFee,
+    TacGasPrice,
     TACSimulationParams,
     TACSimulationResult,
     TransactionLinkerWithOperationId,
@@ -44,6 +46,7 @@ import { DEFAULT_DELAY } from './Consts';
 import { NoopLogger } from './Logger';
 import { OperationTracker } from './OperationTracker';
 import { Simulator } from './Simulator';
+import { TacExplorerClient } from './TacExplorerClient';
 import { TACTransactionManager } from './TACTransactionManager';
 import { TONTransactionManager } from './TONTransactionManager';
 import { TonTxFinalizer } from './TxFinalizer';
@@ -51,6 +54,7 @@ import { getBouncedAddress, mapAssetsToTonAssets, normalizeAssets } from './Util
 export class TacSdk implements ITacSDK {
     readonly config: IConfiguration;
     readonly operationTracker: IOperationTracker;
+    readonly explorerClient: ITacExplorerClient;
     private readonly simulator: ISimulator;
     private readonly tonTransactionManager: ITONTransactionManager;
     private readonly tacTransactionManager: ITACTransactionManager;
@@ -61,12 +65,14 @@ export class TacSdk implements ITacSDK {
         tonTransactionManager: ITONTransactionManager,
         tacTransactionManager: ITACTransactionManager,
         operationTracker: IOperationTracker,
+        explorerClient: ITacExplorerClient,
     ) {
         this.config = config;
         this.simulator = simulator;
         this.tonTransactionManager = tonTransactionManager;
         this.tacTransactionManager = tacTransactionManager;
         this.operationTracker = operationTracker;
+        this.explorerClient = explorerClient;
     }
 
     static async create(sdkParams: SDKParams, logger: ILogger = new NoopLogger()): Promise<TacSdk> {
@@ -99,6 +105,7 @@ export class TacSdk implements ITacSDK {
         );
 
         const operationTracker = new OperationTracker(network, config.liteSequencerEndpoints);
+        const explorerClient = new TacExplorerClient(artifacts.TAC_EXPLORER_API_ENDPOINT);
         const simulator = new Simulator(config, operationTracker, logger);
         const txFinalizer =
             sdkParams.TONParams?.txFinalizer ?? new TonTxFinalizer(config.TONParams.contractOpener, logger);
@@ -111,7 +118,14 @@ export class TacSdk implements ITacSDK {
         );
         const tacTransactionManager = new TACTransactionManager(config, operationTracker, logger);
 
-        return new TacSdk(config, simulator, tonTransactionManager, tacTransactionManager, operationTracker);
+        return new TacSdk(
+            config,
+            simulator,
+            tonTransactionManager,
+            tacTransactionManager,
+            operationTracker,
+            explorerClient,
+        );
     }
 
     closeConnections(): unknown {
@@ -316,6 +330,10 @@ export class TacSdk implements ITacSDK {
         return this.operationTracker;
     }
 
+    getTacExplorerClient(): ITacExplorerClient {
+        return this.explorerClient;
+    }
+
     async prepareCrossChainTransactionPayload(
         evmProxyMsg: EvmProxyMsg,
         senderAddress: string,
@@ -331,8 +349,8 @@ export class TacSdk implements ITacSDK {
         );
     }
 
-    async getTACGasPrice(): Promise<{ average: number; fast: number; slow: number }> {
-        const response = await this.operationTracker.getTACGasPrice();
+    async getTACGasPrice(): Promise<TacGasPrice> {
+        const response = await this.explorerClient.getTACGasPrice();
         return {
             average: response.gasPrices.average,
             fast: response.gasPrices.fast,